feat(auth): Modularized firebase-admin/auth API entrypoint
