+++ src/auth/tenant-manager.ts
@@ -14,18 +14,134 @@
  * limitations under the License.
  */
 
-import { AuthRequestHandler } from './auth-api-request';
-import { FirebaseApp } from '../app/firebase-app';
-import { TenantAwareAuth } from './auth';
-import { Tenant, TenantServerResponse } from './tenant';
-import { AuthClientErrorCode, FirebaseAuthError } from '../utils/error';
 import * as validator from '../utils/validator';
-import { auth } from './index';
+import { App } from '../app';
+import * as utils from '../utils/index';
+import { AuthClientErrorCode, FirebaseAuthError } from '../utils/error';
 
-import ListTenantsResult = auth.ListTenantsResult;
-import TenantManagerInterface = auth.TenantManager;
-import CreateTenantRequest = auth.CreateTenantRequest;
-import UpdateTenantRequest = auth.UpdateTenantRequest;
+import { BaseAuth, SessionCookieOptions } from './base-auth';
+import { Tenant, TenantServerResponse, CreateTenantRequest, UpdateTenantRequest } from './tenant';
+import { FirebaseTokenGenerator, EmulatedSigner, cryptoSignerFromApp } from './token-generator';
+import {
+  AuthRequestHandler, TenantAwareAuthRequestHandler, useEmulator,
+} from './auth-api-request';
+import { DecodedIdToken } from './token-verifier';
+
+/**
+ * Interface representing the object returned from a
+ * {@link auth.TenantManager.listTenants `listTenants()`}
+ * operation.
+ * Contains the list of tenants for the current batch and the next page token if available.
+ */
+export interface ListTenantsResult {
+
+  /**
+   * The list of {@link auth.Tenant `Tenant`} objects for the downloaded batch.
+   */
+  tenants: Tenant[];
+
+  /**
+   * The next page token if available. This is needed for the next batch download.
+   */
+  pageToken?: string;
+}
+
+/**
+ * The tenant aware Auth class.
+ */
+export class TenantAwareAuth extends BaseAuth {
+
+  public readonly tenantId: string;
+
+  /**
+   * The TenantAwareAuth class constructor.
+   *
+   * @param {object} app The app that created this tenant.
+   * @param tenantId The corresponding tenant ID.
+   * @constructor
+   * @internal
+   */
+  constructor(app: App, tenantId: string) {
+    const cryptoSigner = useEmulator() ? new EmulatedSigner() : cryptoSignerFromApp(app);
+    const tokenGenerator = new FirebaseTokenGenerator(cryptoSigner, tenantId);
+    super(app, new TenantAwareAuthRequestHandler(app, tenantId), tokenGenerator);
+    utils.addReadonlyGetter(this, 'tenantId', tenantId);
+  }
+
+  /**
+   * Verifies a JWT auth token. Returns a Promise with the tokens claims. Rejects
+   * the promise if the token could not be verified. If checkRevoked is set to true,
+   * verifies if the session corresponding to the ID token was revoked. If the corresponding
+   * user's session was invalidated, an auth/id-token-revoked error is thrown. If not specified
+   * the check is not applied.
+   *
+   * @param {string} idToken The JWT to verify.
+   * @param {boolean=} checkRevoked Whether to check if the ID token is revoked.
+   * @return {Promise<DecodedIdToken>} A Promise that will be fulfilled after a successful
+   *     verification.
+   */
+  public verifyIdToken(idToken: string, checkRevoked = false): Promise<DecodedIdToken> {
+    return super.verifyIdToken(idToken, checkRevoked)
+      .then((decodedClaims) => {
+        // Validate tenant ID.
+        if (decodedClaims.firebase.tenant !== this.tenantId) {
+          throw new FirebaseAuthError(AuthClientErrorCode.MISMATCHING_TENANT_ID);
+        }
+        return decodedClaims;
+      });
+  }
+
+  /**
+   * Creates a new Firebase session cookie with the specified options that can be used for
+   * session management (set as a server side session cookie with custom cookie policy).
+   * The session cookie JWT will have the same payload claims as the provided ID token.
+   *
+   * @param {string} idToken The Firebase ID token to exchange for a session cookie.
+   * @param {SessionCookieOptions} sessionCookieOptions The session cookie options which includes
+   *     custom session duration.
+   *
+   * @return {Promise<string>} A promise that resolves on success with the created session cookie.
+   */
+  public createSessionCookie(
+    idToken: string, sessionCookieOptions: SessionCookieOptions): Promise<string> {
+    // Validate arguments before processing.
+    if (!validator.isNonEmptyString(idToken)) {
+      return Promise.reject(new FirebaseAuthError(AuthClientErrorCode.INVALID_ID_TOKEN));
+    }
+    if (!validator.isNonNullObject(sessionCookieOptions) ||
+        !validator.isNumber(sessionCookieOptions.expiresIn)) {
+      return Promise.reject(new FirebaseAuthError(AuthClientErrorCode.INVALID_SESSION_COOKIE_DURATION));
+    }
+    // This will verify the ID token and then match the tenant ID before creating the session cookie.
+    return this.verifyIdToken(idToken)
+      .then(() => {
+        return super.createSessionCookie(idToken, sessionCookieOptions);
+      });
+  }
+
+  /**
+   * Verifies a Firebase session cookie. Returns a Promise with the tokens claims. Rejects
+   * the promise if the token could not be verified. If checkRevoked is set to true,
+   * verifies if the session corresponding to the session cookie was revoked. If the corresponding
+   * user's session was invalidated, an auth/session-cookie-revoked error is thrown. If not
+   * specified the check is not performed.
+   *
+   * @param {string} sessionCookie The session cookie to verify.
+   * @param {boolean=} checkRevoked Whether to check if the session cookie is revoked.
+   * @return {Promise<DecodedIdToken>} A Promise that will be fulfilled after a successful
+   *     verification.
+   */
+  public verifySessionCookie(
+    sessionCookie: string, checkRevoked = false): Promise<DecodedIdToken> {
+    return super.verifySessionCookie(sessionCookie, checkRevoked)
+      .then((decodedClaims) => {
+        if (decodedClaims.firebase.tenant !== this.tenantId) {
+          throw new FirebaseAuthError(AuthClientErrorCode.MISMATCHING_TENANT_ID);
+        }
+        return decodedClaims;
+      });
+  }
+}
 
 /**
  * Data structure used to help manage tenant related operations.
@@ -34,15 +150,19 @@ import UpdateTenantRequest = auth.UpdateTenantRequest;
  * - Getting a TenantAwareAuth instance for running Auth related operations (user mgmt, provider config mgmt, etc)
  *   in the context of a specified tenant.
  */
-export class TenantManager implements TenantManagerInterface {
+export class TenantManager {
   private readonly authRequestHandler: AuthRequestHandler;
   private readonly tenantsMap: {[key: string]: TenantAwareAuth};
 
   /**
    * Initializes a TenantManager instance for a specified FirebaseApp.
+   *
    * @param app The app for this TenantManager instance.
+   *
+   * @constructor
+   * @internal
    */
-  constructor(private readonly app: FirebaseApp) {
+  constructor(private readonly app: App) {
     this.authRequestHandler = new AuthRequestHandler(app);
     this.tenantsMap = {};
   }
