+++ test/unit/auth/user-record.spec.ts
@@ -21,9 +21,11 @@ import * as chaiAsPromised from 'chai-as-promised';
 
 import { deepCopy } from '../../../src/utils/deep-copy';
 import {
-  UserInfo, UserMetadata, UserRecord, GetAccountInfoUserResponse, ProviderUserInfoResponse,
-  MultiFactor, PhoneMultiFactorInfo, MultiFactorInfo, MultiFactorInfoResponse,
+  GetAccountInfoUserResponse, ProviderUserInfoResponse, MultiFactorInfoResponse,
 } from '../../../src/auth/user-record';
+import {
+  UserInfo, UserMetadata, UserRecord, MultiFactorSettings, MultiFactorInfo, PhoneMultiFactorInfo,
+} from '../../../src/auth/index';
 
 
 chai.should();
@@ -395,7 +397,7 @@ describe('MultiFactorInfo', () => {
   });
 });
 
-describe('MultiFactor', () => {
+describe('MultiFactorSettings', () => {
   const serverResponse = {
     localId: 'uid123',
     mfaInfo: [
@@ -440,18 +442,18 @@ describe('MultiFactor', () => {
   describe('constructor', () => {
     it('should throw when a non object is provided', () => {
       expect(() =>  {
-        return new MultiFactor(undefined as any);
+        return new MultiFactorSettings(undefined as any);
       }).to.throw('INTERNAL ASSERT FAILED: Invalid multi-factor response');
     });
 
     it('should populate an empty enrolledFactors array when given an empty object', () => {
-      const multiFactor = new MultiFactor({} as any);
+      const multiFactor = new MultiFactorSettings({} as any);
 
       expect(multiFactor.enrolledFactors.length).to.equal(0);
     });
 
     it('should populate expected enrolledFactors', () => {
-      const multiFactor = new MultiFactor(serverResponse);
+      const multiFactor = new MultiFactorSettings(serverResponse);
 
       expect(multiFactor.enrolledFactors.length).to.equal(2);
       expect(multiFactor.enrolledFactors[0]).to.deep.equal(expectedMultiFactorInfo[0]);
@@ -461,7 +463,7 @@ describe('MultiFactor', () => {
 
   describe('getter', () => {
     it('should throw when modifying readonly enrolledFactors property', () => {
-      const multiFactor = new MultiFactor(serverResponse);
+      const multiFactor = new MultiFactorSettings(serverResponse);
 
       expect(() => {
         (multiFactor as any).enrolledFactors = [
@@ -471,7 +473,7 @@ describe('MultiFactor', () => {
     });
 
     it('should throw when modifying readonly enrolledFactors internals', () => {
-      const multiFactor = new MultiFactor(serverResponse);
+      const multiFactor = new MultiFactorSettings(serverResponse);
 
       expect(() => {
         (multiFactor.enrolledFactors as any)[0] = new PhoneMultiFactorInfo({
@@ -486,7 +488,7 @@ describe('MultiFactor', () => {
 
   describe('toJSON', () => {
     it('should return expected JSON object when given an empty response', () => {
-      const multiFactor = new MultiFactor({} as any);
+      const multiFactor = new MultiFactorSettings({} as any);
 
       expect(multiFactor.toJSON()).to.deep.equal({
         enrolledFactors: [],
@@ -494,7 +496,7 @@ describe('MultiFactor', () => {
     });
 
     it('should return expected JSON object when given a populated response', () => {
-      const multiFactor = new MultiFactor(serverResponse);
+      const multiFactor = new MultiFactorSettings(serverResponse);
 
       expect(multiFactor.toJSON()).to.deep.equal({
         enrolledFactors: [
@@ -971,7 +973,7 @@ describe('UserRecord', () => {
     });
 
     it('should return expected multiFactor', () => {
-      const multiFactor = new MultiFactor({
+      const multiFactor = new MultiFactorSettings({
         localId: 'uid123',
         mfaInfo: [
           {
@@ -1001,7 +1003,7 @@ describe('UserRecord', () => {
 
     it('should throw when modifying readonly multiFactor property', () => {
       expect(() => {
-        (userRecord as any).multiFactor = new MultiFactor({
+        (userRecord as any).multiFactor = new MultiFactorSettings({
           localId: 'uid123',
           mfaInfo: [{
             mfaEnrollmentId: 'enrollmentId3',
